feat(page): 메인 화면의 내용을 중첩 라우팅으로 관리
