+++ src/App.tsx
@@ -1,10 +1,12 @@
 import { Routes, Route } from 'react-router-dom';
-import { MainPage } from '@/pages';
+import { MainPage, CongestionPage } from '@/pages';
 
 function App() {
   return (
     <Routes>
-      <Route path='/' element={<MainPage />} />
+      <Route path='/' element={<MainPage />}>
+        <Route path='' element={<CongestionPage />} />
+      </Route>
     </Routes>
   );
 }
