+++ src/components/common/Gnb/Menu.tsx
@@ -1,4 +1,5 @@
 import SvgIcon from '@/components/common/Icon/SvgIcon';
+import { Link } from 'react-router-dom';
 import tw from 'twin.macro';
 
 const Menu = ({
@@ -8,6 +9,7 @@ const Menu = ({
   strokeWidth = 1.2,
   styles,
   title,
+  link,
 }: {
   icon: string;
   width: number;
@@ -15,13 +17,24 @@ const Menu = ({
   strokeWidth?: number;
   title?: string;
   styles?: any;
+  link?: string;
 }) => {
+  const Icon = () => (
+    <div css={tw`flex justify-center items-center`}>
+      <div css={[styles]}>{SvgIcon[icon](width, height, strokeWidth)}</div>
+      <p css={tw`p-10`}>{title}</p>
+    </div>
+  );
+
   return (
     <div css={[tw`flex justify-between items-center w-full`]}>
-      <div css={tw`flex justify-center items-center`}>
-        <div css={[styles]}>{SvgIcon[icon](width, height, strokeWidth)}</div>
-        <p css={tw`p-10`}>{title}</p>
-      </div>
+      {link ? (
+        <Link to={link}>
+          <Icon />
+        </Link>
+      ) : (
+        <Icon />
+      )}
     </div>
   );
 };
